function_app/requirements-functions.txt
anthropic>=0.49.0
azure-core>=1.33.0
azure-functions>=1.23.0
azure-identity>=1.21.0
azure-keyvault-secrets>=4.9.0
azure-mgmt-resource>=23.3.0
azure-storage-blob>=12.25.1
email-validator>=2.2.0
feedparser>=6.0.11
flask-babel>=4.0.0
flask>=3.1.0
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
langdetect>=1.0.9
markdown>=3.8
matplotlib>=3.10.1
newspaper3k>=0.2.8
nltk>=3.9.1
openai>=1.75.0
pandas>=2.2.1
psycopg2-binary>=2.9.10
pygments>=2.19.1
python-dateutil>=2.9.0.post0
python-dotenv>=1.0.1
python-slugify>=8.0.4
pytrends>=4.9.2
pyyaml>=6.0.1
requests>=2.32.3
seaborn>=0.13.2
slack-sdk>=3.35.0
spacy>=3.8.5
textblob>=0.19.0
tiktoken>=0.9.0
trafilatura>=2.0.0
wordcloud>=1.9.4
sendgrid>=6.11.0
flake8>=7.2.0